Simple_Neurite_Tracer: when loading SWC files, offer a useful options dialog
In practice there seem to be a variety of interpretations of the SWC file
specification, and sometimes one is given files that need to be loaded at a
particular offset into an image, or flipped in one axis, etc.

This dialog allows you to correct the most common problem (image co-ordinates
rather than world co-ordinates being used) with the top checkbox, and also
apply an arbitrary scale and offset.