Simple_Neurite_Tracer: warn about points outside the image volume, rather than trying to fix them
Previously we detected that some SWC files had their co-ordinates
reflected in one axis or another and try to fix it, but this was
a fudge - files that validly had points outside the image volume
would have been misrepresented.  This change means that now we just
count the points outside the image volume, and put a warning in the
log if any are found.  The better way of dealing with these files
would be to use the new "scale" option in the SWC import dialog with
a scale of -1, so the warning message suggests adjusting these
options.