Simple_Neurite_Tracer: some SWC files specify 0 radius points - change these on import
This change sets any points with 0 radius to half the minimum voxel
separation on import, which at least means that something is rendered.
(This isn't really the right fix - it doesn't distinguish between
files where all radii are set to 0 and those where that's the case
for only some points.  However, as an interim measure, hopefully it
won't cause problems.)