Simple_Neurite_Tracer: grab the keystrokes the plugin needs, and pass on other
This fixes a long-standing problem where the plugin was not correctly
overriding ImageJ's default key presses for the key shortcuts it uses.

I've fixed this by creating a new KeyListener, called
QueueJumpingKeyListener which can have a list of fallback
keylisteners added to it - it will try those if it's not
interested in the keypress until the event is consumed.

There's also a helper method, setAsFirstKeyListener, which
will will remove and save all the key listeners from a
component, add them to a new QueueJumpingKeyListener and
then set the QueueJumpingKeyListener as the only key listener
for the component.

This is necessary because there's no guaranteed order in which
the key listeners are called, so if you want to be called first
you have to make sure that you're the only key listener.  There's
some discussion of that here:

   http://stackoverflow.com/questions/5274172/is-the-order-in-which-keylisteners-will-be-called-guaranteed