Simple_Neurite_Tracer: the 'm' key "clicks" at the median maximum point perpendicular to the plane
Thank you to Uygar Sumbul for suggesting this feature.  The idea is
to support a workflow where one overlays the MIP projection
of the stack with 10% opacity.  This means that it's possible to
keep slice set at the soma, click in the soma, and then just press
'm' with the mouse over what appears to be the end of a dendrite
in the overlay.  Then you can repeat without having to scroll
through the stack so much.