Simple_Neurite_Tracer: support communication with TrakEM2 or other plugins
This commit adds support for adding listeners to Simple Neurite Tracer,
so that messages can be sent to them.  This also adds a "Send to TrakEM2"
menu option, which causes a SEND_TO_TRAKEM2 message to be sent to all
listeners.  (In response, the listener can get information from SNT.)
This also adds a QUIT message, for detecting the user leaving the plugin.