Simple Neurite Tracer: use Fiji's JUnit naming convention
Classes implementing unit tests for JUnit are supposed to end in "Test" or
start with "Test". In Fiji, we decided to use only the former convention.

This will most likely fix the breakage reported by Jacey Jenkins, but for
a different reason: a legacy-imglib1 artifact with work-arounds for a
broken Bio-Formats 5.0.0 POM hierarchy was just deployed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>