create notebooks for performance testing and convex hull
- create one jupyter notebook to quantify performance of converting between Java and Python data types
- create another jupyter notebook to perform the convex hull calculation and rendering
- removed the initial commit which is less performant 

The convex hull of a neuron reconstruction is one way of quantifying how much space a cell takes up within a compartment(s). 
A future addition could be to compute the convex hulls of 2+ cells and measure the amount of overlap,
in order to assess spatial groupings of terminal node clusters.