Simplify "Fork at node" shortcut.
Previous we used Ctrl, but because Ctrl on Mac functions as popup menu trigger, we were using Alt on mac. We also want to support built-in shortcuts triggered using Ctrl/Cmd, so for simplification we will use Alt on all platforms. This automatically fixes "Fork point" label not being updated in real-time described in #48