add multiple_mesh_intersection notebook
This commit adds an IPython notebook which demonstrates the calculation and rendering of meshes that result from spatial boolean operations, such as intersection and union, between two meshes. 

This may be useful, for example, to quantify a rough estimate of the amount of spatial overlap between two distinct dendritic trees.

Requirements:
-pyimagej
-trimesh + Blender
-numpy

Tested on Windows 10 64-bit