Viewer3D: Do not allow frame to jump between screens on linux
On Linux when using multi-displays the frame would appear in between screens, and start rendering extremely slow somehow. With awesome window manager, it would jump in between screens. This fix centers the frame on the primary display, which seems to bypass all this issues. While at it, it also reduces the verbosity of debug mode messages, that were slowing things down